Changed DestroyAfterTime script to check if the game object is on screen before destroying, changed name of script
