+++ Assets/Prefabs/Asteroid.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 5814717036466881330}
   - component: {fileID: 8881510278450501433}
   - component: {fileID: 3087972472837476525}
-  - component: {fileID: 1708341211821687673}
+  - component: {fileID: 3424410005269634327}
   m_Layer: 0
   m_Name: Asteroid
   m_TagString: Obstacle
@@ -131,12 +131,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fb4939d7e31eea04892871c0fbc8f9ab, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  tf: {fileID: 8657655460236217139}
   body: {fileID: 6609337249155862677}
   minScale: 0.3
   maxScale: 1.5
   changeMass: 1
---- !u!114 &1708341211821687673
+--- !u!114 &3424410005269634327
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -145,7 +144,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 3199362265423141486}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bf0751d1abfd4a45b4cb896de9d01cb, type: 3}
+  m_Script: {fileID: 11500000, guid: fa801952a200cb34eabec05ea6b24c71, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   timeToLive: 15
