+++ Assets/Prefabs/Bird.prefab
@@ -13,9 +13,9 @@ GameObject:
   - component: {fileID: 6780121967607222692}
   - component: {fileID: 7137176313116615387}
   - component: {fileID: 6727194292044465682}
-  - component: {fileID: 4909964578888700895}
   - component: {fileID: 486122271932694736}
   - component: {fileID: 2034419137310108267}
+  - component: {fileID: 2475371463945810206}
   m_Layer: 0
   m_Name: Bird
   m_TagString: Obstacle
@@ -120,19 +120,6 @@ MonoBehaviour:
   StartVelocityY: 0
   leftOrRight: 0
   percentageOfSpawnBounds: 0.5
---- !u!114 &4909964578888700895
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7001849492976900931}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bf0751d1abfd4a45b4cb896de9d01cb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  timeToLive: 100
 --- !u!114 &486122271932694736
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -163,6 +150,19 @@ MonoBehaviour:
   minScale: 0.1
   maxScale: 0.2
   changeMass: 0
+--- !u!114 &2475371463945810206
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7001849492976900931}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fa801952a200cb34eabec05ea6b24c71, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  timeToLive: 10
 --- !u!1001 &7610135424270664595
 PrefabInstance:
   m_ObjectHideFlags: 0
