+++ Assets/Prefabs/FallingRock.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 5814717036466881330}
   - component: {fileID: 8881510278450501433}
   - component: {fileID: 3087972472837476525}
-  - component: {fileID: 888381267915208969}
+  - component: {fileID: 4229513799280990240}
   m_Layer: 0
   m_Name: FallingRock
   m_TagString: Obstacle
@@ -135,7 +135,7 @@ MonoBehaviour:
   minScale: 0.1
   maxScale: 0.5
   changeMass: 1
---- !u!114 &888381267915208969
+--- !u!114 &4229513799280990240
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -144,10 +144,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 3199362265423141486}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bf0751d1abfd4a45b4cb896de9d01cb, type: 3}
+  m_Script: {fileID: 11500000, guid: fa801952a200cb34eabec05ea6b24c71, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  timeToLive: 10
+  timeToLive: 5
 --- !u!1001 &4574594432676789377
 PrefabInstance:
   m_ObjectHideFlags: 0
