+++ Assets/Prefabs/PaperAirPlane.prefab
@@ -13,8 +13,8 @@ GameObject:
   - component: {fileID: 6780121967607222692}
   - component: {fileID: 6727194292044465682}
   - component: {fileID: 2689787620967252856}
-  - component: {fileID: 4909964578888700895}
   - component: {fileID: 6821191503183044842}
+  - component: {fileID: 1520653579693893621}
   m_Layer: 0
   m_Name: PaperAirPlane
   m_TagString: Obstacle
@@ -121,19 +121,6 @@ MonoBehaviour:
   timeDescending: 2
   descendingAcceleration: 0.1
   ascendingAcceleration: 0.2
---- !u!114 &4909964578888700895
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7001849492976900931}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bf0751d1abfd4a45b4cb896de9d01cb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  timeToLive: 100
 --- !u!114 &6821191503183044842
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -148,6 +135,19 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   tf: {fileID: 4759502860409365753}
   rb: {fileID: 6780121967607222692}
+--- !u!114 &1520653579693893621
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7001849492976900931}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fa801952a200cb34eabec05ea6b24c71, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  timeToLive: 10
 --- !u!1001 &4147937219332713056
 PrefabInstance:
   m_ObjectHideFlags: 0
