+++ Assets/Prefabs/Projectile.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 8193641594476504203}
   - component: {fileID: 4041241860910258351}
   - component: {fileID: 2504093771395316230}
-  - component: {fileID: 3503334190783364971}
   m_Layer: 0
   m_Name: Projectile
   m_TagString: Obstacle
@@ -137,16 +136,3 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &3503334190783364971
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8696507923222920497}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bf0751d1abfd4a45b4cb896de9d01cb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  timeToLive: 4
