+++ Assets/Prefabs/Twig.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 5814717036466881330}
   - component: {fileID: 8881510278450501433}
   - component: {fileID: 3087972472837476525}
-  - component: {fileID: 8333335195194874934}
+  - component: {fileID: -7354270966530453246}
   m_Layer: 0
   m_Name: Twig
   m_TagString: Obstacle
@@ -135,7 +135,7 @@ MonoBehaviour:
   minScale: 0.2
   maxScale: 0.4
   changeMass: 1
---- !u!114 &8333335195194874934
+--- !u!114 &-7354270966530453246
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -144,10 +144,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 3199362265423141486}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3bf0751d1abfd4a45b4cb896de9d01cb, type: 3}
+  m_Script: {fileID: 11500000, guid: fa801952a200cb34eabec05ea6b24c71, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  timeToLive: 10
+  timeToLive: 5
 --- !u!1001 &2330568915450578900
 PrefabInstance:
   m_ObjectHideFlags: 0
