+++ Assets/Scripts/DestoryAfterTimeOffScreen.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fa801952a200cb34eabec05ea6b24c71
