--- Assets/Scripts/DestroyAfterTime.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3bf0751d1abfd4a45b4cb896de9d01cb
